refactor: added a sub-folder for table row component
