+++ app/component/table_row/table_row_component.html.erb
