+++ app/views/groups/show.html.erb
@@ -77,7 +77,7 @@
     <% end %>
     <tbody class="groups-expanded-table">
       <% @group.assignments.includes(projects: :grade).each do |assignment| %>
-        <%= render(TableRowComponent.new(assignment, current_user, {
+        <%= render(TableRow::TableRowComponent.new(assignment, current_user, {
           group: @group,
           policy: policy(@group),
           policy_assignment: policy(assignment)
